Saving credit history of every card and take credit to card

+++ src/main/java/com/home/model/primitive/Number.java
@@ -1,16 +1,16 @@
 package com.home.model.primitive;
 
 public class Number {
-    private Integer number;
+    private Double number;
 
     public Number() {
     }
 
-    public Integer getNumber() {
+    public Double getNumber() {
         return number;
     }
 
-    public void setNumber(Integer number) {
+    public void setNumber(Double number) {
         this.number = number;
     }
